+++ src/Properties.java
@@ -1,4 +1,4 @@
 public class Properties {
-    public static int BUS_INTERVAL_MEAN = 1200;
-    public static int RIDER_INTERVAL_MEAN = 30;
+    public static double BUS_INTERVAL_MEAN = 1200.0;
+    public static double RIDER_INTERVAL_MEAN = 30.0;
 }
